src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via book-agent (pyproject.toml)
filetype==1.2.0
    # via langchain-google-genai
google-ai-generativelanguage==0.6.18
    # via langchain-google-genai
google-api-core==2.25.1
    # via google-ai-generativelanguage
google-auth==2.40.3
    # via
    #   book-agent (pyproject.toml)
    #   google-ai-generativelanguage
    #   google-api-core
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.74.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   book-agent (pyproject.toml)
    #   langgraph-sdk
    #   langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.3.74
    # via
    #   book-agent (pyproject.toml)
    #   langchain-google-genai
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==2.1.9
    # via book-agent (pyproject.toml)
langgraph==0.6.6
    # via book-agent (pyproject.toml)
langgraph-checkpoint==2.1.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.6.4
    # via langgraph
langgraph-sdk==0.2.2
    # via langgraph
langsmith==0.4.15
    # via langchain-core
orjson==3.11.2
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   pytest
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==6.32.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via
    #   book-agent (pyproject.toml)
    #   fastapi
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via pytest
pytest==8.4.1
    # via
    #   book-agent (pyproject.toml)
    #   pytest-asyncio
pytest-asyncio==1.1.0
    # via book-agent (pyproject.toml)
python-dotenv==1.1.1
    # via uvicorn
pyyaml==6.0.2
    # via
    #   langchain-core
    #   uvicorn
requests==2.32.5
    # via
    #   google-api-core
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
tenacity==9.1.2
    # via langchain-core
typing-extensions==4.14.1
    # via
    #   anyio
    #   fastapi
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via book-agent (pyproject.toml)
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
xxhash==3.5.0
    # via langgraph
zstandard==0.24.0
    # via langsmith
